Implement special tokens extraction for transformers (#41)
The transformers extractor did not support extraction of representations
for special tokens like `[CLS]`, `[SEP]`, `[PAD]` etc. This commit adds
an additional parameter to also extract these along with the original
tokens.

Detailed commit history:

* Implement special tokens extraction for transformers

The transformers extractor did not support extraction of representations
for special tokens like `[CLS]`, `[SEP]`, `[PAD]` etc. This commit adds
an additional parameter to also extract these along with the original
tokens.

* Add check for special/dropped token adjacency

When a special token follows a token that was completely dropped by
the tokenizer, or vice-versa, we do not have enough information to
rebuild the original input and now explictly raise an exception.

* Refactor transformers tests for clarity

* Add tests for long inputs

* Add command line option in extractor

* Format code

* Add documentation for extraction functions